+++ bindings/nodejs/bin/install-gpu.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 'use strict'
 
-require('../index.js')
+require('../gpu-installer.js')
     .installGpu()
     .catch((err) => {
         console.error(err.message || err)